PR 변경사항 수정 및 반응형 디자인
